[MOV] evaluation_test: merge tests files in one
Before this commit, the tests of evaluation were in two differents files,
which not make sense.
